fix: removeMyLecture 최상단으로 이동
